+++ atl_core/src/atl_includes.h
