+++ atl_core/src/model/ChildrenResult.h
