+++ atl_core/src/model/Model.h
