+++ atl_core/src/model/Result.h
