+++ atl_core/src/model/StackMap.h
