+++ atl_core/src/service/AtlService.cpp
