+++ atl_core/src/service/AtlService.h
