+++ atl_core/src/service/assertion/IsEqualTo.h
