+++ atl_core/src/service/assertion/ToStringTemplate.h
