+++ atl_core/src/template.h
