+++ atl_core/src/testBuilder/TestClassBuilder.cpp
