+++ atl_core/src/testBuilder/TestClassBuilder.h
