+++ example/atl_app/src/MyTests.h
