+++ example/libs/aLib/aLibTest/src/MyModule.h
