+++ example/libs/aLib/src/Item.cpp
@@ -1,0 +1,15 @@
+#pragma once
+#include "Item.h"
+
+bool Item::operator==(Item &rhs)
+{
+	return name == rhs.name && value == rhs.value;
+}
+
+std::string (*ItemtoString)(Item it) =
+	[](Item it) -> std::string {
+	return std::string("Name : ")
+		.append(it.name)
+		.append(" and value : ")
+		.append(std::to_string(it.value));
+};
